moving test into common module
